1、超级会员逻辑 2、奖学金逻辑 3、轮播图点击逻辑
git-svn-id: svn://120.24.217.103/repository/outsource_projects/wu_xb/android/shop/XiaoeShop@253957 d1851369-bd24-4d56-8f42-d5f936fd0249

+++ common/src/main/java/xiaoe/com/common/entitys/WrItem.java
@@ -1,9 +1,11 @@
 package xiaoe.com.common.entitys;
 
+import java.math.BigDecimal;
+
 public class WrItem {
 
     private String wrTitle; // 提现记录 item 标题
-    private int wrMoney; // 提现记录 item 金额
+    private BigDecimal wrMoney; // 提现记录 item 金额
     private String wrTime; // 提现记录 item 时间
     private String wrState; // 提现记录 item 状态
 
@@ -11,7 +13,7 @@ public class WrItem {
         this.wrTitle = wrTitle;
     }
 
-    public void setWrMoney(int wrMoney) {
+    public void setWrMoney(BigDecimal wrMoney) {
         this.wrMoney = wrMoney;
     }
 
@@ -28,7 +30,7 @@ public class WrItem {
         return wrTitle;
     }
 
-    public int getWrMoney() {
+    public BigDecimal getWrMoney() {
         return wrMoney;
     }
